+++ doc/changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+V0.0.7
+- Added a trigger for acquisition container type (SYSTEM trigger when running gear rules)
+
+V0.0.6
+- Updated Docker image to 0.0.6
+- Improved output file organization
+    - if session level output as normal
+    - catch for fiducial file being missing if run on cpu, not gpu
+
 ## 0.0.0
 ```
 NJB: 4/11/2024